Add scrollup function and update the like and comments without refreshing the page.
