+++ package.json
@@ -7,7 +7,8 @@
     "dev": "webpack --mode development",
     "start": "webpack serve --open",
     "build": "webpack",
-    "test": "jest --coverage"
+    "test": "jest --coverage",
+    "deploy": "gh-pages -d dist/"
   },
   "repository": {
     "type": "git",
@@ -45,6 +46,7 @@
   "dependencies": {
     "@popperjs/core": "^2.11.6",
     "bootstrap": "^4.6.0",
+    "gh-pages": "^4.0.0",
     "i": "^0.3.7",
     "jest-environment-jsdom": "^29.0.3",
     "jquery": "^3.6.1",
