+++ src/components/addNewComment.js
@@ -1,9 +1,9 @@
 import formData from './formData.js';
 import formValidation from './formValidation.js';
+import updateComments from './updateComments.js';
 
 const addNewComment = async (e) => {
   e.preventDefault();
-  const itemId = e.currentTarget.parentElement.parentElement.parentElement.parentElement.childNodes[1].childNodes[1].childNodes[9].childNodes[1].id.split('-')[1];
   const mealDiv = e.currentTarget.parentElement.parentElement.parentElement.parentElement;
   const commentDiv = document.createElement('div');
   commentDiv.classList.add('mealdiv');
@@ -24,17 +24,9 @@ const addNewComment = async (e) => {
   e.currentTarget.parentElement.parentElement.parentElement.parentElement.childNodes[1].childNodes[1].childNodes[5].style.justifyContent = 'space-evenly';
   e.currentTarget.parentElement.parentElement.parentElement.parentElement.childNodes[1].childNodes[1].childNodes[7].childNodes[1].style.float = 'none';
   e.currentTarget.parentElement.parentElement.parentElement.parentElement.childNodes[1].childNodes[1].childNodes[7].style.textAlign = 'center';
+
+  const itemId = e.currentTarget.parentElement.parentElement.parentElement.parentElement.childNodes[1].childNodes[1].childNodes[9].childNodes[1].id.split('-')[1];
+  updateComments(itemId);
   document.querySelector('#comment-submit').addEventListener('click', formData);
-  const response = await fetch(`https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/lZ1jpmvfp72iObG11Ico/comments?item_id=${itemId}`);
-  const json = await response.json();
-  document.querySelector('.comment-div').innerHTML = `<h3>Comments(${json.length})</h3>`;
-  json.forEach((element) => {
-    const commentDetails = document.createElement('div');
-    commentDetails.innerHTML = `
-    <small>${element.creation_date}</small>
-    <small>${element.username}:</small>
-    <small>${element.comment}</small>`;
-    document.querySelector('.comment-div').appendChild(commentDetails);
-  });
 };
 export default addNewComment;
