+++ src/index.html
@@ -12,6 +12,7 @@
     </div>
     <div id="meal">
     </div>
+    <div class="scroll-up"></div>
     <div id="footerContainer">
     </div>
 </body>
