+++ src/index.js
@@ -3,6 +3,7 @@ import './style.css';
 import header from './components/header.js';
 import loadAll from './components/loadAll.js';
 import footer from './components/footer.js';
+import scrollUp from './components/scrollUp.js';
 
 // PRINT HEADER
 header();
@@ -12,3 +13,10 @@ loadAll();
 
 // PRINT THE FOOTER
 footer();
+
+// SCROLLUP up
+scrollUp();
+document.querySelector('#scroll').addEventListener('click', (e) => {
+  e.preventDefault();
+  window.scrollTo({ top: 0, behavior: 'smooth' });
+});
